REFACTOR: Move Configuration to Configuration folder
