+++ src/Arriba/Arriba.Server/Controllers/OAuthController.cs
@@ -1,4 +1,5 @@
 ﻿using Arriba.Configuration;
+using Arriba.Server.Configuration;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using System;
